REST 组件

+++ sample-web/src/main/java/com/yewei/sample/service/UserService.java
@@ -12,10 +12,10 @@ import java.util.List;
 public interface UserService {
 
     //这里只写了个add方法，其它的直接在控制层调用Dao层，正常开发流程都应该写在Service层
-    Integer add(UserModel user)throws Exception;
+    Long add(UserModel user)throws Exception;
     List<UserResponse> listAllUsers(UserQueryRequest query);
     PageResult<UserResponse> findUsersByPage(UserQueryParam query);
-    UserResponse findById(long id) throws Exception;
-    Boolean deleteById(long id)throws Exception;
+    UserResponse findById(Long id) throws Exception;
+    Boolean deleteById(Long id)throws Exception;
     Boolean updateById(UserUpdateRequest request)throws Exception;
 }
